testing king move logic
