+++ src/main/java/engine/EntityCreator.java
@@ -3,7 +3,7 @@ package main.java.engine;
 import main.java.game.entity.piece.Checker;
 import main.java.game.entity.piece.KingChecker;
 import main.java.game.utils.AssetManager;
-import main.java.game.utils.GameBoardPiece;
+import main.java.game.entity.GameBoardPiece;
 
 import java.awt.Image;
 
@@ -16,14 +16,15 @@ public class EntityCreator {
 
     public GameBoardPiece createPiece(String name, int x, int y) {
         String pieceName = name.substring(5);
-        switch (pieceName) {
-            case "Checker":
-                return createChecker(name, x, y);
-            case "CheckerKing":
-                return createKingChecker(name, x, y);
-            default:
-                throw new IllegalArgumentException("Invalid piece name: " + name + " tested as: " + pieceName);
-        }
+//        switch (pieceName) {
+//            case "Checker":
+//                return createChecker(name, x, y);
+//            case "CheckerKing":
+//                return createKingChecker(name, x, y);
+//            default:
+//                throw new IllegalArgumentException("Invalid piece name: " + name + " tested as: " + pieceName);
+//        }
+        return createKingChecker(name + "King", x, y);
     }
 
     private Checker createChecker(String name, int x, int y) {
